Implement basic persitent collapsible comments
During RedditComments.init(), build a tree ADT of the comments and use that to
drive the UI changes.  An array of collapsed comment IDs is calculated and
saved to localstorage on each collapse and expand operation.  During initial
load, collapsed and hidden classes are applied to the relevant elements to
effect the state.

Next steps: improve collapsed display by including commenter, timestamp, etc.
Consider how changes to the collapsed threads are indicated on visible parents.