Change behavior of hnes-new/hnes-new-parent when parent is also new
When looking at a previously viewed item, a person's eye follows the
trail of hnes-new-parent nodes looking for the hnes-new nodes. But with
the current behavior some of the hnes-new-parent nodes are themselves
new nodes and it's easy to miss.

This change makes sure that any new node is colored with hnes-new while
the previous nodes on the path get colored with hnes-new-parent like
before.